% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec50-workflow.R
\name{fit_quality}
\alias{fit_quality}
\alias{fit_failures}
\title{Inspect EC50 fit quality and failures}
\usage{
fit_quality(x)

fit_failures(x)
}
\arguments{
\item{x}{An object returned by [estimate_EC50()] or [ec50_multimodel()].}
}
\value{
A data frame.
}
\description{
`fit_quality()` returns group-level quality information for successful fits.
`fit_failures()` returns failed group/model combinations as data.
}
\examples{
data(multi_isolate)
fit <- estimate_EC50(
  growth ~ dose,
  data = subset(multi_isolate, isolate \%in\% 1:3 & fungicida == "Fungicide A"),
  isolate_col = "isolate",
  strata_col = "field",
  fct = drc::LL.3()
)
fit_quality(fit)
fit_failures(fit)

}
